requirements.txt
vanna[gemini,bigquery,chromadb]
vertexai
python-dotenv
sqlparse
google-cloud-bigquery[pandas]